+++ CHANGELOG.md
@@ -43,6 +43,16 @@ cheatcodes = { git = "https://github.com/foundry-rs/starknet-foundry", tag = "0.
     - `target_name_for_package`
     - `compilation_unit_for_package`
 
+### Cast
+
+#### Added
+
+- Added `--class-hash` flag to account create/deploy, allowing for custom openzeppelin account contract class hash 
+
+#### Changed
+
+-
+
 ## [0.3.0] - 2023-08-02
 
 ### Forge